+++ kit/dapp/src/components/asset-designer/shared-form.ts
@@ -1,4 +1,5 @@
 import { AssetDesignerStepSchema } from "@/components/asset-designer/steps";
+import { isRequiredFieldForSchema } from "@/lib/utils/schema-field";
 import { BondSchema } from "@/orpc/routes/token/routes/mutations/create/helpers/create-handlers/bond.create.schema";
 import { FundSchema } from "@/orpc/routes/token/routes/mutations/create/helpers/create-handlers/fund.create.schema";
 import { TokenBaseSchema } from "@/orpc/routes/token/routes/mutations/create/helpers/token.base-create.schema";
@@ -16,31 +17,7 @@ export type AssetDesignerFormInputData = z.input<
 >;
 
 export const isRequiredField = (field: keyof AssetDesignerFormInputData) => {
-  //   const schema = AssetDesignerFormSchema.pick({
-  //     [field]: true,
-  //   });
-
-  //   const isOptional = schema.safeParse({
-  //     [field]: undefined,
-  //   }).success;
-
-  // return !isOptional;
   return isRequiredFieldForSchema(AssetDesignerFormSchema, field);
-};
-
-export const isRequiredFieldForSchema = <TSchema extends z.ZodObject>(
-  schema: TSchema,
-  field: keyof z.input<TSchema>
-): boolean => {
-  const fieldSchema = schema.pick({
-    [field]: true as const,
-  });
-
-  const isOptional = fieldSchema.safeParse({
-    [field]: undefined,
-  }).success;
-
-  return !isOptional;
 };
 
 export const assetDesignerFormOptions = formOptions({
